requirements.txt
altair==5.5.0
anyio==3.7.1
appdirs==1.4.4
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
bs4==0.0.2
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cssselect==1.3.0
cycler==0.12.1
dateparser==1.2.2
deep-translator==1.11.4
demjson3==3.0.6
facebook-scraper==0.2.59
fake-useragent==2.2.0
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
h11==0.16.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langdetect==1.0.9
lxml==6.0.2
lxml_html_clean==0.4.3
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
narwhals==2.5.0
networkx==3.5
nltk==3.9.1
numpy==2.3.3
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
parse==1.20.2
pillow==11.3.0
plotly==6.3.0
proto-plus==1.26.1
protobuf==6.32.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydeck==0.9.1
pyee==11.1.1
pyparsing==3.2.5
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-html==0.10.0
rfc3986==1.5.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
Sastrawi==1.0.1
selenium==4.36.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@614d4c2029a62d348ca56598f87c425966aaec66
sortedcontainers==2.4.0
soupsieve==2.8
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
textblob==0.19.0
tokenizers==0.22.1
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.0
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.26.20
w3lib==2.3.1
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==10.4
wordcloud==1.9.4
wsproto==1.2.0
zipp==3.23.0
